Renaming "MicroFocus" with "OpenText"
