+++ dotnet/AdapterSdkSchema.sln
@@ -12,7 +12,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sepg", "sepg", "{C2BEB224-7
 		pom.xml = pom.xml
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MicroFocus.FAS.AdapterSdkSchema", "src\MicroFocus.FAS.AdapterSdkSchema\MicroFocus.FAS.AdapterSdkSchema.csproj", "{F88720DC-490A-42D3-9FF7-FFD05CB4529A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenText.FAS.AdapterSdkSchema", "src\OpenText.FAS.AdapterSdkSchema\OpenText.FAS.AdapterSdkSchema.csproj", "{F88720DC-490A-42D3-9FF7-FFD05CB4529A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MicroFocus.FAS.AdapterSdkSchema.Tests", "tests\MicroFocus.FAS.AdapterSdkSchema.Tests\MicroFocus.FAS.AdapterSdkSchema.Tests.csproj", "{B4F57D4F-B5B9-431A-8AF5-3927D9068C67}"
 EndProject
@@ -20,7 +20,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MicroFocus.FAS.AdapterSdkSc
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MicroFocus.FAS.AdapterSdkSchema.Generator.Tests", "tests\MicroFocus.FAS.AdapterSdkSchema.Generator.Tests\MicroFocus.FAS.AdapterSdkSchema.Generator.Tests.csproj", "{348DFA4B-7117-4534-967A-F4D58E3AEDC5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MicroFocus.FAS.AdapterSdkSchema.JsonBuilders.System", "src\MicroFocus.FAS.AdapterSdkSchema.JsonBuilders.System\MicroFocus.FAS.AdapterSdkSchema.JsonBuilders.System.csproj", "{15CFB8B2-B957-4CBD-B084-FDCB73023893}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenText.FAS.AdapterSdkSchema.JsonBuilders.System", "src\OpenText.FAS.AdapterSdkSchema.JsonBuilders.System\OpenText.FAS.AdapterSdkSchema.JsonBuilders.System.csproj", "{15CFB8B2-B957-4CBD-B084-FDCB73023893}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
