+++ dotnet/src/OpenText.FAS.AdapterSdkSchema/IStructuredSubfield.cs
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace MicroFocus.FAS.AdapterSdkSchema
+namespace OpenText.FAS.AdapterSdkSchema
 {
     public interface IStructuredSubfield : IStructuredField, ISubfield
     {
