+++ dotnet/src/OpenText.FAS.AdapterSdkSchema/SchemaObjectBuilders/BooleanExtension.cs
@@ -13,9 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace MicroFocus.FAS.AdapterSdkSchema
+namespace OpenText.FAS.AdapterSdkSchema.SchemaObjectBuilders
 {
-    public interface IStructuredField : IField
+    internal static class BooleanExtension
     {
+        public static string ToLowerCaseString(this bool that)
+        {
+            return that ? "true" : "false";
+        }
     }
 }
