+++ dotnet/src/OpenText.FAS.AdapterSdkSchema/SubfieldImpl.cs
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace MicroFocus.FAS.AdapterSdkSchema
+namespace OpenText.FAS.AdapterSdkSchema
 {
     internal class SubfieldImpl : FieldImpl, ISubfield
     {
